requirements.txt
kenlm
PyYAML
